% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scores}
\alias{scores}
\title{Computing the Cell State Deviations (CSDs) and the Cell Anomalousness Scores (CASs)}
\usage{
scores(feature.mat, anchor.frac = 0.15, n.cores = 1, verbose = T, K = 30)
}
\arguments{
\item{feature.mat}{a matrix containing the feature vectors for each cells, each row represents one feature}

\item{anchor.frac}{the fraction of anchor cells, 0.15 by default}

\item{n.cores}{threads called for computation, 1 by default}

\item{verbose}{logical, whether report the computation progress}

\item{K}{the number of the nearest neighbors in computing the CASs}
}
\value{
list, containing CSDs and CASs
}
\description{
Computing the Cell State Deviations (CSDs) and the Cell Anomalousness Scores (CASs)
}
