% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualization of CSDs and CASs}
\usage{
visualize(feature.mat, scores, cell.labels, size = 0.25)
}
\arguments{
\item{feature.mat}{a matrix containing the feature vectors for each cells, each row represents one feature}

\item{scores}{a list returned by scores() function, containing the CSDs and CASs computed}

\item{cell.labels}{vector, the annotations, cell types or clusters of individual cells}

\item{size}{the size of points, o.25 by default}
}
\value{
a list of three t-SNE visualization plots containing the gradient plot of CSDs and CASs and the visualization of cells with labels
}
\description{
Visualization of CSDs and CASs
}
